reworked server for nodemailer, updated react router & folders, updated concurrently script
