+++ server/config/mailerConfig.js
@@ -1,0 +1,6 @@
+require('dotenv').config()
+
+module.exports = {
+    USER: process.env.emailUser,
+    PASS: process.env.emailPass
+}
