--- src/config/config.js
@@ -1,6 +1,0 @@
-require('dotenv').config()
-
-module.exports = {
-  USER: process.env.emailUser,
-  PASS: process.env.emailPass
-}
