+++ src/index.js
@@ -4,15 +4,16 @@ import './index.css';
 import Home from './pages/Home';
 import Services from './pages/Services';
 import SampleService from './pages/SampleService';
-import NavBar from './components/NavBar';
-import App from './App'
+// import NavBar from './components/NavBar';
+import ContactUsPage from './pages/ContactUs'
+import App from './App.js'
 import * as serviceWorker from './serviceWorker';
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 
 ReactDOM.render(
   <React.StrictMode>
 
-    <NavBar />
+
     <Router>
       <Switch>
         <Route exact path="/" component={Home} />
@@ -20,6 +21,7 @@ ReactDOM.render(
         <Route exact path="/services" component={Services} />
         <Route exact path="/sampleservice" component={SampleService} />
         <Route exact path="/admin" component={App} />
+        <Route exact path="/contactus" component={ContactUsPage} />
       </Switch>
     </Router>
